refactor: rename _etrace_input_data to _input_data and update imports
