+++ .gitignore
@@ -209,3 +209,4 @@ cython_debug/
 /.claudegitignore
 /.codex
 /AGENTS.md
+/research/
