+++ braintrace/_etrace_graph_executor.py
@@ -40,7 +40,7 @@ from typing import Dict, Any
 import brainstate
 
 from braintrace._etrace_compiler import ETraceGraph, compile_etrace_graph
-from ._etrace_input_data import get_single_step_data
+from ._input_data import get_single_step_data
 from ._typing import Path
 
 __all__ = [
