+++ braintrace/_etrace_vjp/base.py
@@ -22,7 +22,7 @@ import jax.numpy as jnp
 import saiunit as u
 
 from braintrace._etrace_algorithms import ETraceAlgorithm
-from braintrace._etrace_input_data import has_multistep_data
+from braintrace._input_data import has_multistep_data
 from braintrace._state_managment import assign_state_values_v2
 from braintrace._typing import (
     PyTree,
